agentic_hierarchical_rag/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LLM APIs
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.1.0

# Clustering and analysis
scipy>=1.10.0
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
nltk>=3.8.0
spacy>=3.6.0
tiktoken>=0.5.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Evaluation and benchmarking
datasets>=2.14.0
evaluate>=0.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0